fixed new addon structure
